Update crc to 5.0.0
As 5.0.0 now properly advertises typing information by a `py.typed`
marker in the package, this change drops the suppressions of the lack of
type information complaint by Pylance/Pyright and mypy type checkers.

As typing information is not a runtime dependency, this version
requirement is only relevant for the 'development' requirements set.